update HgModel and rendering path to handle enormous models. we can render at least 5,000,000 vertices constructing 10,000,000 triangles

+++ Mercury3/src/renderer/RenderData.h
@@ -68,7 +68,7 @@ public:
 	uint32_t index_count;
 
 	uint32_t vbo_offset;
-	uint16_t vertex_count;
+	uint32_t vertex_count;
 
 	BlendMode blendMode;
 	RenderFlags renderFlags;
